requirements.txt
openai>=2.0.0
python-dotenv>=1.0.0
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scipy>=1.11.0
adjusttext>=0.8.0
